+++ src/primitives.rs
@@ -4,5 +4,5 @@ pub mod sphere;
 use crate::vector3::Vector3;
 
 pub trait SignedDistance: Sync {
-    fn distance_from(&self, position: Vector3) -> f32;
+    fn distance_from(&self, position: Vector3) -> f64;
 }
